requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"